Trying swapping a script src url
Since we are using static rather than a server I think we need to jiggle some things? I'm not sure how this was set up before. Must be some kind of build system, but in that case, what was the build doing?
I really shouldn't do this kind of experimentation on github... This will be the only edit I make until the weekend when I can use my own computer.